refactor: change feng auth handler place
